+++ docs/pages/components/message/api/message.js
@@ -73,6 +73,13 @@ export default [
                 type: 'String',
                 values: '—',
                 default: '—'
+            },
+            {
+                name: '<code>aria-close-label</code>',
+                description: 'Label for the close button, to be read by accessibility screenreaders.',
+                type: 'String',
+                values: '—',
+                default: '—'
             }
         ],
         events: [