+++ docs/pages/components/message/examples/ExTypes.vue
@@ -1,22 +1,22 @@
 <template>
     <section>
-        <b-message title="Default">
+        <b-message title="Default" aria-close-label="Close message">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
         </b-message>
 
-        <b-message title="Danger" type="is-danger">
+        <b-message title="Danger" type="is-danger" aria-close-label="Close message">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
         </b-message>
 
-        <b-message title="Success" type="is-success">
+        <b-message title="Success" type="is-success" aria-close-label="Close message">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
         </b-message>
 
-        <b-message title="Info" type="is-info">
+        <b-message title="Info" type="is-info" aria-close-label="Close message">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
         </b-message>
 
-        <b-message title="Warning" type="is-warning">
+        <b-message title="Warning" type="is-warning" aria-close-label="Close message">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
         </b-message>
     </section>