+++ src/components/message/Message.vue
@@ -11,6 +11,7 @@
                     type="button"
                     class="delete"
                     @click="close"
+                    :aria-label="ariaCloseLabel"
                 />
             </header>
             <section class="message-body">
@@ -38,6 +39,9 @@
     export default {
         name: 'BMessage',
         mixins: [MessageMixin],
+        props: {
+            ariaCloseLabel: String
+        },
         data() {
             return {
                 newIconSize: this.iconSize || this.size || 'is-large'